I've refactored the database code and file structure. Here's what I did:
- I moved `scripts/strainr-db.py` to `src/strainr/build_db.py`.
- I merged `src/strainr/kmer_database.py` (class `StrainKmerDb`) and `src/strainr/database.py` (class `StrainKmerDatabase`) into `src/strainr/database.py`.
  - The merged class is now named `StrainKmerDatabase`.
  - It's based on the more comprehensive `StrainKmerDb`, keeping its features like k-mer length inference, detailed loading, and example usage block.
- I removed the original `src/strainr/kmer_database.py`.
- I updated all relevant import statements across your codebase to reflect these changes.
- I verified the merged `StrainKmerDatabase` class through its example usage block.
