+++ src/strainr/__init__.py
@@ -10,7 +10,7 @@ __version__ = "0.1.0"
 
 # Core classes and functions for easier access
 from .analyze import ClassificationAnalyzer
-from .kmer_database import StrainKmerDb  # Updated to consolidated class name
+from .database import StrainKmerDatabase # Updated to consolidated class name
 from .output import AbundanceCalculator
 from .parameter_config import process_arguments
 from .running import main
@@ -18,7 +18,7 @@ from .sequence import GenomicSequence, extract_kmers_from_sequence
 from .utils import get_canonical_kmer, open_file_transparently
 
 __all__ = [
-    "StrainKmerDb",  # Updated to consolidated class name
+    "StrainKmerDatabase",  # Updated to consolidated class name
     "ClassificationAnalyzer",
     "AbundanceCalculator",
     "GenomicSequence",
